clean up JSON responses

+++ app/controllers/tasks_controller.rb
@@ -1,6 +1,6 @@
 class TasksController < ApplicationController
     respond_to :json
-    # before_action :authenticate_user!
+    before_action :authenticate_user!
     
     def index
         @project = Project.find_by(id: params['projectId'])
@@ -9,17 +9,21 @@ class TasksController < ApplicationController
     
     def show
         @task = Task.find_by(id: params['id'])
-        @comments = @task.comments.collect do |comment| 
-            {
-                author: comment.author.username,
-                author_id: comment.author.id,
-                content: comment.content,
-                created_at: comment.created_at,
-                task_id: comment.task_id,
-                id: comment.id
-            }
+        if @task
+            @comments = @task.comments.collect do |comment| 
+                {
+                    author: comment.author.username,
+                    author_id: comment.author.id,
+                    content: comment.content,
+                    created_at: comment.created_at,
+                    task_id: comment.task_id,
+                    id: comment.id
+                }
+            end
+            render json: {task: @task, comments: @comments}, status: 200
+        else
+            render json: {message: 'Something went wrong!'}, status: 500
         end
-        render json: {task: @task, comments: @comments}
     end
 
     def create
@@ -35,10 +39,14 @@ class TasksController < ApplicationController
 
     def destroy
         @task = Task.find_by(id: params[:id])
-        if @task.destroy
-            render json: {message: 'Successfully Deleted', status: 200}
+        if current_user == @task.creator || current_user == @task.project.team.leader
+            if @task.destroy
+                render json: {message: 'Successfully Deleted'}, status: 200
+            else
+                render json: {message: 'Something Went Wrong!'}, status: 500
+            end
         else
-            render json: {message: 'Something Went Wrong!', status: 401}
+            render json: {message: 'You are not authorized to do that'}, status: 500
         end
     end
 end
